+++ src/components/BottomSheet/ModalBottomSheet.tsx
@@ -79,7 +79,6 @@ const styles = StyleSheet.create({
     borderTopLeftRadius: 24,
     borderTopRightRadius: 24,
     backgroundColor: '#FFFFFF',
-    justifyContent: 'space-between',
     padding: 25,
   },
 });
